Use a wrapper class for URLRequest
This allows us to avoid making the URLRequest itself conform to
Codable, which could theoretically clash with other frameworks
doing the same thing. This also makes the request code more closely
mirror the response code within MockDuck, and gives us a place to add
other request data if we ever need it.