Allow MockDuck to affect other URLSessions
* Add a function for preparing any URLSessionConfiguration for use by
MockDuck. This can be called on any URLSessionConfiguration object to
inject MockURLProtocol as a valid protocolClass, enabling MockDuck to
handle any requests made by the URLSession configured by that object.

* Swizzle URLSessionConfiguration.default and
URLSessionConfiguration.ephemeral to inject MockURLProtocol as a valid
protocolClass. This allows MockDuck to work with other frameworks
(like Alamofire) with no additional configuration.