Improve MockRequestResponse handling
This commit now creates a single MockRequestResponse which lives for
the lifecycle of a MockDuck request. This allows this object to cache
and store the normalized request value, making it so that each
request is only normalized a single time by the MockDuckDelegate.
This change allows users of the framework to do clever things like
increment an integer each time a certain URL is normalized, append that
integer to the path, and thus return different data each time that URL
is requested.

In doing this, a lot of logic in MockSerializable and SerializationUtils
was moved into MockRequestResponse and RequestResponseCommonProtocol. I
believe that overall this makes the code a bit easier to follow and
should also make it easier to edit this code in the future.